sync

+++ util/Image.php
@@ -5,13 +5,14 @@ class Image
 {
 	private $path = "Midia"; 
 	private $tam; 
-	private $errors; 
+	public  $errors; 
 
 	function __construct(){
 		$this->tam = 1028 * 1028 * 5;
-		$this->errors = array('Houve um problema no uploado de arquivo',
-							  'O arquivo é muito grande', 
-							  'Erro no servidor.'
+		$this->errors = array('houve um problema no upload de arquivo.',
+							  'o arquivo é muito grande.', 
+							  'servidor com problemas.', 
+							  'envie um arquivo de imagem PNG ou JPG.'
 							 );  
 	}
 
@@ -21,15 +22,19 @@ class Image
 			return 0; 
 		
 		if($_FILES[$name]['size'] > $this->tam )
-			return 2; 
+			return 1; 
 		
-		$extension = explode('.',$_FILES[$name]['name']); 
+		$extension = strtolower(explode('.',$_FILES[$name]['name'])[1]); 
+
+		if(strcmp($extension,'png') != 0 && strcmp($extension,'jpg') != 0 )
+			return 3; 
+
 		$nomeFinal = md5(time().$_FILES[$name]['name']).'.'.$extension[1]; 
 
 		if(move_uploaded_file($_FILES[$name]['tmp_name'],PATH.$this->path.DS.$nomeFinal))
 			return PATH.$this->path.DS.$nomeFinal; 
 		else
-			return 3; 
+			return 2; 
 	}
 
 
@@ -37,7 +42,7 @@ class Image
 		$result = self::saveImg($name);
 
 		if(is_int($result))
-			return $this->errors[$result]; 
+			return $result; 
 
 		$wide = WideImage::load($result); 
 
@@ -45,6 +50,26 @@ class Image
 		$nomeFinal = md5(time().$_FILES[$name]['name'].'A-D').'.'.$extension[1]; 
 
 		$wide->crop($x,$y,$x2,$y2)->resize(128)->saveToFile(PATH.$this->path.DS.$nomeFinal,$compress); 
+		//$wide->crop($x,$y,$x2,$y2)->resize(128)->output('jpg', $compress);
+
+		unlink($result); 
+		return $nomeFinal; 
+	}
+
+	public function saveCropLandspace($name,$x=0,$y=0,$x2=100,$y2=100,$compress=50){
+		$result = self::saveImg($name);
+
+		if(is_int($result))
+			return $result; 
+
+		$wide = WideImage::load($result); 
+
+		$extension = explode('.',$_FILES[$name]['name']); 
+		$nomeFinal = md5(time().$_FILES[$name]['name'].'A-D').'.'.$extension[1]; 
+
+		var_dump($extension); 
+		
+		$wide->crop($x,$y,$x2,$y2)->resize(1000,400)->saveToFile(PATH.$this->path.DS.$nomeFinal,$compress); 
 		//$wide->crop($x,$y,$x2,$y2)->resize(128)->output('jpg', $compress);
 
 		unlink($result);
